move validation to a separate folder to be later extracted into its own package
